// generated/src/app/api-models/stage-auth-addtl-info-pwk.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class StageAuthAddtlInfoPwk {
  batchId: string;
  transactionId: number;
  seqStageAuthAddtlInfoPwk: number;
  seqStageAuthProc: number;
  paperworkLevelInd: string;
  rptTypeCode: string;
  transCode: string;
  ctrlNo: string;
  description: string;
  insertDatetime: Date;
}